% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drg_aveBoxplot.R
\name{drg_aveBoxplot}
\alias{drg_aveBoxplot}
\title{drg_aveBoxplot}
\usage{
drg_aveBoxplot(data, varx)
}
\arguments{
\item{varx}{}
}
\value{
boxplot
}
\description{
This function makes the boxplot of payments by DRG code.
There are options for the function to make the boxplot for either
the average Medicare payments or the average total payment.
}
\examples{
drg_aveBoxplot(drg, 'Average.Medicare.Payments')
drg_aveBoxplot(drg, 'Average.Total.Payments')

}
