% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_statstable.R
\name{DRG_stat}
\alias{DRG_stat}
\title{Statistics by DRG Code Function}
\usage{
DRG_stat(df, stat)
}
\arguments{
\item{df}{This is the dataframe of DRG data to be used (DRG_data.csv in this lab)}

\item{stat}{This is the statistic you would like to generate a table of: mean (mean), median (median), or standard deviation (sd)}
}
\value{
An HTML table of the mean, median or SD of the payment variables over different DRG codes
}
\description{
Statistics by DRG Code Function
}
\examples{
DRGdata <- read.csv("DRG_data.csv")
DRG_stat(DRGdata, stat = "sd")

}
